backend/requirements.txt
firebase-admin
fastapi
uvicorn
apscheduler
google-generativeai
pillow
python-multipart
